Safari-611.1.10/LayoutTests/media/modern-media-controls/audio/audio-controls-buttons-expected.txt

Testing the <audio> controls buttons.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Making AirPlay routes available.
AirPlay routes became available.
PASS mediaController.controls.rightContainer.children.length is 2
PASS mediaController.controls.rightContainer.children[0] is mediaController.controls.muteButton
PASS mediaController.controls.rightContainer.children[1] is mediaController.controls.airplayButton
PASS accessibilityController.accessibleElementById('audio-controls').description is "AXDescription: Audio Controls"

PASS successfullyParsed is true

TEST COMPLETE
